exercise-07/requirements.txt
google-cloud-bigquery==3.11.4
pandas==2.1.4
db-dtypes==1.1.1